+++ backend/src/main/java/db/pojo/Training.java
@@ -7,13 +7,13 @@ public interface Training extends Serializable {
 
     void setUser(UserData user);
 
-    int getId();
+    long getId();
 
-    void setId(int id);
+    void setId(long id);
 
-    int getUserId();
+    long getUserId();
 
-    void setUserId(int userId);
+    void setUserId(long userId);
 
     String getCreateDate();
